+++ pytorch_lightning/callbacks/model_checkpoint.py
@@ -176,8 +176,8 @@ class ModelCheckpoint(Callback):
 
         if not isinstance(current, torch.Tensor):
             rank_zero_warn(
-                f'{current} is supposed to be a torch.Tensor. Saving checkpoint may not work correctly. '
-                f'HINT: check the value of {self.monitor} in your validation loop', RuntimeWarning
+                f'{current} is supposed to be a `torch.Tensor`. Saving checkpoint may not work correctly.'
+                f' HINT: check the value of {self.monitor} in your validation loop', RuntimeWarning
             )
             current = torch.tensor(current)
 
@@ -294,9 +294,11 @@ class ModelCheckpoint(Callback):
 
             if not isinstance(current, torch.Tensor):
                 rank_zero_warn(
-                    f'The metric you returned {current} must be a Torch.Tensor instance, checkpoint not saved '
-                    f'HINT: what is the value of {self.monitor} in validation_epoch_end()?', RuntimeWarning
+                    f'The metric you returned {current} must be a `torch.Tensor` instance, checkpoint not saved'
+                    f' HINT: what is the value of {self.monitor} in validation_epoch_end()?', RuntimeWarning
                 )
+                if current is not None:
+                    current = torch.tensor(current)
 
             if current is None:
                 rank_zero_warn(